+++ CHANGES.rst
@@ -5,6 +5,8 @@ Version 8.3.x
 
 Unreleased
 
+-   Fix Fish shell completion errors when option help text contains newlines.
+    :issue:`3043`
 -   Don't discard pager arguments by correctly using ``subprocess.Popen``. :issue:`3039`
     :pr:`3055`
 -   Replace ``Sentinel.UNSET`` default values by ``None`` as they're passed through